+++ DownloadableCodeProjects/LP2_implement_interfaces/Interfaces_M1/Solution/ImplementInterfaces/IPerson.cs
@@ -6,15 +6,4 @@ namespace ImplementInterfaces
         int Age { get; set; }
         void DisplayInfo();
     }
-
-    public class Student : IPerson
-    {
-        public string Name { get; set; } = string.Empty;
-        public int Age { get; set; } = 0;
-
-        public void DisplayInfo()
-        {
-            Console.WriteLine($"Student Name: {Name}, Age: {Age}");
-        }
-    }
 }
